Namespace change for condition checkers
