+++ src/HttPlaceholder.Application.Tests/StubExecution/ConditionChecking/FormValueConditionCheckerFacts.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using HttPlaceholder.Application.Interfaces.Http;
-using HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations;
+using HttPlaceholder.Application.StubExecution.ConditionChecking;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 using Microsoft.Extensions.Primitives;
