+++ src/HttPlaceholder.Application.Tests/StubExecution/ConditionChecking/HeaderConditionCheckerFacts.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using HttPlaceholder.Application.Interfaces.Http;
-using HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations;
+using HttPlaceholder.Application.StubExecution.ConditionChecking;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
