+++ src/HttPlaceholder.Application.Tests/StubExecution/ConditionChecking/JsonPathConditionCheckerFacts.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using HttPlaceholder.Application.Interfaces.Http;
-using HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations;
+using HttPlaceholder.Application.StubExecution.ConditionChecking;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
