+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/BasicAuthenticationConditionChecker.cs
@@ -4,7 +4,7 @@ using HttPlaceholder.Application.Interfaces.Http;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class BasicAuthenticationConditionChecker : IConditionChecker
     {
