+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/FormValueConditionChecker.cs
@@ -6,7 +6,7 @@ using HttPlaceholder.Common.Utilities;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class FormValueConditionChecker : IConditionChecker
     {
