+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/IsHttpsConditionChecker.cs
@@ -2,7 +2,7 @@
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class IsHttpsConditionChecker : IConditionChecker
     {
