+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/MethodConditionChecker.cs
@@ -1,15 +1,15 @@
-﻿using HttPlaceholder.Application.Interfaces.Http;
-using HttPlaceholder.Common.Utilities;
+﻿using System;
+using HttPlaceholder.Application.Interfaces.Http;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
-    public class FullPathConditionChecker : IConditionChecker
+    public class MethodConditionChecker : IConditionChecker
     {
         private readonly IHttpContextService _httpContextService;
 
-        public FullPathConditionChecker(IHttpContextService httpContextService)
+        public MethodConditionChecker(IHttpContextService httpContextService)
         {
             _httpContextService = httpContextService;
         }
@@ -17,27 +17,27 @@ namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementat
         public ConditionCheckResultModel Validate(StubModel stub)
         {
             var result = new ConditionCheckResultModel();
-            var fullPathCondition = stub.Conditions?.Url?.FullPath;
-            if (string.IsNullOrEmpty(fullPathCondition))
+            var methodCondition = stub.Conditions?.Method;
+            if (string.IsNullOrEmpty(methodCondition))
             {
                 return result;
             }
 
-            var path = _httpContextService.FullPath;
-            if (StringHelper.IsRegexMatchOrSubstring(path, fullPathCondition))
+            var method = _httpContextService.Method;
+            if (string.Equals(methodCondition, method, StringComparison.OrdinalIgnoreCase))
             {
                 // The path matches the provided regex. Add the stub ID to the resulting list.
                 result.ConditionValidation = ConditionValidationType.Valid;
             }
             else
             {
-                result.Log = $"Condition '{fullPathCondition}' did not pass for request.";
+                result.Log = $"Condition '{methodCondition}' did not pass for request.";
                 result.ConditionValidation = ConditionValidationType.Invalid;
             }
 
             return result;
         }
 
-        public int Priority => 9;
+        public int Priority => 10;
     }
 }
