+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/PathConditionChecker.cs
@@ -3,7 +3,7 @@ using HttPlaceholder.Common.Utilities;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class PathConditionChecker : IConditionChecker
     {
