+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/QueryStringConditionChecker.cs
@@ -4,7 +4,7 @@ using HttPlaceholder.Common.Utilities;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class QueryStringConditionChecker : IConditionChecker
     {
