+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/ScenarioExactHitCounterConditionChecker.cs
@@ -1,13 +1,13 @@
 ﻿using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
-    public class ScenarioMaxHitCounterConditionChecker : IConditionChecker
+    public class ScenarioExactHitCounterConditionChecker : IConditionChecker
     {
         private readonly IScenarioService _scenarioService;
 
-        public ScenarioMaxHitCounterConditionChecker(IScenarioService scenarioService)
+        public ScenarioExactHitCounterConditionChecker(IScenarioService scenarioService)
         {
             _scenarioService = scenarioService;
         }
@@ -15,8 +15,8 @@ namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementat
         public ConditionCheckResultModel Validate(StubModel stub)
         {
             var result = new ConditionCheckResultModel();
-            var maxHits = stub.Conditions?.Scenario?.MaxHits;
-            if (maxHits == null)
+            var exactHits = stub.Conditions?.Scenario?.ExactHits;
+            if (exactHits == null)
             {
                 return result;
             }
@@ -29,13 +29,13 @@ namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementat
                 result.Log = "No hit count could be found.";
                 result.ConditionValidation = ConditionValidationType.Invalid;
             }
-            else if (actualHitCount >= maxHits)
+            else if (actualHitCount != exactHits)
             {
                 result.Log =
-                    $"Scenario '{scenario}' should have less than '{maxHits}' hits, but '{actualHitCount}' hits were counted.";
+                    $"Scenario '{scenario}' should have exactly '{exactHits}' hits, but '{actualHitCount}' hits were counted.";
                 result.ConditionValidation = ConditionValidationType.Invalid;
             }
-            else if (actualHitCount < maxHits)
+            else
             {
                 result.ConditionValidation = ConditionValidationType.Valid;
             }
