+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/ScenarioMinHitCounterConditionChecker.cs
@@ -1,13 +1,13 @@
 ﻿using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
-    public class ScenarioExactHitCounterConditionChecker : IConditionChecker
+    public class ScenarioMinHitCounterConditionChecker : IConditionChecker
     {
         private readonly IScenarioService _scenarioService;
 
-        public ScenarioExactHitCounterConditionChecker(IScenarioService scenarioService)
+        public ScenarioMinHitCounterConditionChecker(IScenarioService scenarioService)
         {
             _scenarioService = scenarioService;
         }
@@ -15,8 +15,8 @@ namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementat
         public ConditionCheckResultModel Validate(StubModel stub)
         {
             var result = new ConditionCheckResultModel();
-            var exactHits = stub.Conditions?.Scenario?.ExactHits;
-            if (exactHits == null)
+            var minHits = stub.Conditions?.Scenario?.MinHits;
+            if (minHits == null)
             {
                 return result;
             }
@@ -29,13 +29,13 @@ namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementat
                 result.Log = "No hit count could be found.";
                 result.ConditionValidation = ConditionValidationType.Invalid;
             }
-            else if (actualHitCount != exactHits)
+            else if (actualHitCount < minHits)
             {
                 result.Log =
-                    $"Scenario '{scenario}' should have exactly '{exactHits}' hits, but '{actualHitCount}' hits were counted.";
+                    $"Scenario '{scenario}' should have at least '{minHits}' hits, but only '{actualHitCount}' hits were counted.";
                 result.ConditionValidation = ConditionValidationType.Invalid;
             }
-            else
+            else if (actualHitCount >= minHits)
             {
                 result.ConditionValidation = ConditionValidationType.Valid;
             }
