+++ src/HttPlaceholder.Application/StubExecution/ConditionChecking/XPathConditionChecker.cs
@@ -5,7 +5,7 @@ using HttPlaceholder.Common.Utilities;
 using HttPlaceholder.Domain;
 using HttPlaceholder.Domain.Enums;
 
-namespace HttPlaceholder.Application.StubExecution.ConditionChecking.Implementations
+namespace HttPlaceholder.Application.StubExecution.ConditionChecking
 {
     public class XPathConditionChecker : IConditionChecker
     {
